add minimal projects structure to bind components to clusters
This patch introduces a minimal Project definition and a ProjectBuilder
definition to build them.

The purpose is to introduce a concept, a Project, for the purpose of
binding components to clusters.  Once we have this defined and a way to
build concrete values of them, we can use it to feed helm chart values
using varying aspects of the Project fields.

To start we organize around regions and zones, largely because that's
how `#Regions` is defined and the `regions` structure is laid out.

This allows us to lookup the clusters in a regional zone, then bind the
component to each.

Next, we'll work on defining how to actually feed values to the charts.

The previous platform metadata and podinfo examples have been disabled,
we'll eventually handle the same configuration in the #ProjectBuilder by
filling out the labels and parameters and what not in the Component
defined by the Project, built by the ProjectBuilder.

Here's what it looks like:

```
holos show platform
```

```yaml
apiVersion: v1alpha5
kind: Platform
metadata:
  name: default
spec:
  components:
    - name: podinfo
      path: components/podinfo
    - name: podinfo
      path: components/podinfo
    - name: podinfo
      path: components/podinfo
``