remove #Project.clusters field and podinfo
The #Project.clusters field was copied from the OIS definition of a
project and isn't needed in this context.

The podinfo component was copied from a previous branch and isn't
needed, focus instead on the provided components.

```
holos show platform
```

Still not useful, but is wired up.

```yaml
apiVersion: v1alpha5
kind: Platform
metadata:
  name: default
spec:
  components:
    - name: argocd
      path: components/argocd
    - name: argocd
      path: components/argocd
    - name: argocd
      path: components/argocd
```