define and add ComponentsByScopes data structure
Previously there was no clear place to add a component to all clusters
of a given scope, which is necessary to address a business requirement
such as, "Manage ArgoCD on all management clusters, but not internal or
customer clusters."

I'm not sure this is an actual business requirement, but if we imagine
it is the ComponetsByScopes structure provides this capability.

The ArgoCD component is listed, but doesn't have any cluster, region, or
zonal specific data.  It only indicates argocd is associated with the
mgmt scope and the path of the component is at components/argocd.

The regions struct is where specific components are assoicated with a
scope, region, and zone.

The ProjectBuilder is responsible for composing the values of a specific
cluster, region, zone, and scope into a component.  For example, each
component results in build plans that write to an output path.  The
output path contains the cluster and project name, so the ProjectBuilder
composes these values into the Component.

Once the project is built, its componets roll up into the Platform spec.

Here's the result.  We still haven't achieved the Scenario 1 goal of
mapping helm values to components, but the structures are shaping up.

The work remaining at this point is to:

1. Loop over regions and zones, we're only managing zone dev1a now.
2. Pass key attributes (region, zone, scope) to the component to use as
   lookup keys for the correct helm values.
3. Expand to more components than ArgoCD.

```
holos show platform
```

```yaml
apiVersion: v1alpha5
kind: Platform
metadata:
  name: default
spec:
  components:
    - annotations:
        app.holos.run/description: argocd for project experiment on cluster dev1a-mgmt
      labels:
        app.holos.run/name: argocd
      name: argocd
      parameters:
        outputBaseDir: clusters/dev1a-mgmt/projects/experiment
      path: components/argocd
```

```
❯ holos render platform
rendered argocd for project experiment on cluster dev1a-mgmt in 852.502833ms
rendered platform in 853.110708ms
```