manage argocd in mgmt clusters in all regions and zones
Previously we were only managing ArgoCD on the mgmt scoped cluster in
the dev1a zone.  We assume we want it managed on the mgmt cluster in
each zone, and automatically managed if we add more zones and regions to
the regions struct.

This patch loops over the regions and zones and composes a "setup"
project into each.  Setup is a poor name, but the project is intended to
manage the components we need to initially set things up.  ArgoCD is
usually first when bootstraping a new zone or cluster, so we add it to
this setup project.  Alternative names for the project might be "infra"
or "platform" but these are so generic and overloaded they're probably
not any better.

In any event, we're starting to see how the composition is coming
together.  Here's the result:

```
holos show platform
```

```yaml
apiVersion: v1alpha5
kind: Platform
metadata:
  name: default
spec:
  components:
    - annotations:
        app.holos.run/description: argocd for project setup on cluster dev1a-mgmt
      labels:
        app.holos.run/name: argocd
      name: argocd
      parameters:
        outputBaseDir: clusters/dev1a-mgmt/projects/setup
      path: components/argocd
    - annotations:
        app.holos.run/description: argocd for project setup on cluster us-west-01a-mgmt
      labels:
        app.holos.run/name: argocd
      name: argocd
      parameters:
        outputBaseDir: clusters/us-west-01a-mgmt/projects/setup
      path: components/argocd
    - annotations:
        app.holos.run/description: argocd for project setup on cluster us-west-01b-mgmt
      labels:
        app.holos.run/name: argocd
      name: argocd
      parameters:
        outputBaseDir: clusters/us-west-01b-mgmt/projects/setup
      path: components/argocd
    - annotations:
        app.holos.run/description: argocd for project setup on cluster us-east-02a-mgmt
      labels:
        app.holos.run/name: argocd
      name: argocd
      parameters:
        outputBaseDir: clusters/us-east-02a-mgmt/projects/setup
      path: components/argocd
```

```
holos render platform
```

```
rendered argocd for project setup on cluster us-west-01a-mgmt in 718.473ms
rendered argocd for project setup on cluster us-west-01b-mgmt in 726.895084ms
rendered argocd for project setup on cluster us-east-02a-mgmt in 728.136333ms
rendered argocd for project setup on cluster dev1a-mgmt in 734.350625ms
rendered platform in 734.46475ms
```