add schema for argo-cd chart values
Previously the argo-cd chart is ignoring the values we pass because they
aren't correct and don't match the default values.yaml schema.

This patch imports the default values from the chart as the `#Values`
schema.

The result is the values fail to verify, in the next change we'll relax
the imported #Values definition to accomidate the values we're
providing.

  holos show buildplans

  could not run:
    holos.spec.artifacts.0.generators.0.helm.values.global.tolerations:
    incompatible list lengths (0 and 1) (and 1 more errors) at
    internal/builder/instance.go:66
  holos.spec.artifacts.0.generators.0.helm.values.global.tolerations:
    incompatible list lengths (0 and 1)