demo values based on a given cluster scope use case
Previously we managed the argocd component on clusters in the mgmt scope
but no other cluster scopes.  We have no example demonstrating how a
component can span multiple cluster scopes and still vary by which scope
it resides in for a given cluster.

This commit demonstrates how to pass values that vary based on the
cluster scope to a component deployed to multiple cluster scopes.

The components/cluster-scope-values is meant to represent a component
you might deploy to all of your clusters.  Values are pased to a
ConfigMap just to illustrate getting the data all the way through Holos
to the cluster.

The notable addition with this patch is the introduction of a @tag for
the ClusterScope field.  When holos renders each platform component, it
will pass the current cluster's scope as this tag value.  This behavior
is configured by the parameters field in the platform component list.

For example: (abbreviated)

  holos show platform

  apiVersion: v1alpha5
  kind: Platform
  metadata:
    name: default
  spec:
    components:
      - annotations:
          app.holos.run/description: cluster-scope-values for project setup on cluster dev1a-internal
        labels:
          app.holos.run/name: cluster-scope-values
        name: cluster-scope-values
        parameters:
          ClusterScope: internal
          outputBaseDir: clusters/dev1a-internal/projects/setup
        path: components/cluster-scope-values

The components/cluster-scope-values/ component definition uses the
ClusterScope field as a lookup key into the ValuesByClusterScope
structure, defined back at the root of the repository.

This change produces a large number of manifests, so the next patch adds
those to keep this patch focused on the code.

Result:

  holos render platform

  rendered cluster-scope-values for project setup on cluster us-west-01a-mgmt in 167.185709ms
  rendered cluster-scope-values for project setup on cluster us-west-01a-customer in 168.365958ms
  rendered cluster-scope-values for project setup on cluster dev1a-customer in 170.536542ms
  rendered cluster-scope-values for project setup on cluster us-east-02a-mgmt in 171.014583ms
  rendered cluster-scope-values for project setup on cluster us-west-01b-mgmt in 184.808458ms
  rendered cluster-scope-values for project setup on cluster us-east-02a-internal in 185.883917ms
  rendered cluster-scope-values for project setup on cluster dev1a-internal in 190.868333ms
  rendered cluster-scope-values for project setup on cluster us-west-01a-internal in 201.439458ms
  rendered cluster-scope-values for project setup on cluster us-west-01b-internal in 214.182708ms
  rendered cluster-scope-values for project setup on cluster dev1a-mgmt in 222.472584ms
  rendered cluster-scope-values for project setup on cluster us-west-01b-customer in 229.995ms
  rendered cluster-scope-values for project setup on cluster us-east-02a-customer in 126.595292ms
  rendered cluster-scope-values for project setup on cluster us-east-02a-customer3 in 124.2535ms
  rendered cluster-scope-values for project setup on cluster us-east-02a-customer2 in 124.905667ms
  rendered argocd for project setup on cluster us-west-01b-mgmt in 765.375083ms
  rendered argocd for project setup on cluster dev1a-mgmt in 770.733833ms
  rendered argocd for project setup on cluster us-west-01a-mgmt in 809.842209ms
  rendered argocd for project setup on cluster us-east-02a-mgmt in 708.934125ms
  rendered platform in 876.332041ms